requirements.txt
nemo_toolkit
fastapi
uvicorn
python-multipart
pydantic
torch
numpy
hydra-core
lightning
lhotse
einops
transformers
sentencepiece
pandas
jiwer
librosa
pyannote-audio
webdataset
datasets
editdistance
IPython
pydub
ffmpeg-python
soundfile
cuda-python
